refactor: move logError to error package
