+++ src/errors/index.ts
@@ -1,2 +1,1 @@
-export * from './logError';
 export * from './displayMutationError';
