+++ src/packages/errors/index.ts
@@ -1,1 +1,2 @@
 export * from './transactionErrors';
+export * from './logError';
