+++ src/packages/errors/logError.ts
